bq_test_data_generation/requirements_market_data.txt
# Requirements for generating and uploading stock market data to BigQuery

pandas>=1.5.0
numpy>=1.21.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
pandas-gbq>=0.26.1